backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "chromadb>=1.0.10",
  "django>=5.2.1",
  "djangorestframework>=3.16.0",
  "huey>=2.5.3",
  "openai>=1.82.0",
  "python-dotenv>=1.1.0",
  "redis>=6.1.0",
  "sentence-transformers>=4.1.0",
  "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
  "django-types>=0.20.0",
  "djangorestframework-types>=0.9.0",
  "ruff>=0.11.10",
]
